Handle attributes like "warning" and "alert"
This also makes w53 controllable with warning attributes, but only for top-level
alerts.  This is used to handle the fact that in the stdlib there's a template
file with a top-level alert attribute, and it's included in two different
places, one where that attribute is legal, and one where it is not

A boostrap is needed, and performed in the next commit.